Leave CAPTCHA Recognition/requirements.txt
absl-py==1.3.0
astunparse==1.6.3
blinker==1.8.2
brotlipy==0.7.0
cachetools==5.2.0
click==8.1.7
Flask==3.0.3
flatbuffers==1.12
gast==0.4.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.7.0
keras==2.9.0
Keras-Preprocessing==1.1.2
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
mkl-fft==1.3.6
mkl-service==2.4.0
numpy==1.24.1
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==22.0
Pillow==9.4.0
protobuf==3.19.6
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.11.1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.1.1
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
typing_extensions==4.4.0
Werkzeug==3.0.3
wrapt==1.14.1

